split code base into smaller sections
